separate setup_tar_and_args method
